+++ test/middleware/strictness.test.ts
@@ -9,7 +9,7 @@ describe("Defining a state type is enforced", () => {
 
     const wrapperThatAddsSession = (handler: Handler<{}, {}, {session: "123"}>) => {
         return async (ctx: Context<any>, next: MiddlewareNextFunction) => {
-            ctx.state.session = "123";
+            ctx.data.session = "123";
             handler(ctx as Context<any, {session: "123"}>, next)
         }
     }
@@ -20,9 +20,9 @@ describe("Defining a state type is enforced", () => {
     // This test is only for the type system, so no need to actually check anything here.
     expect(true).toEqual(true);
 
-    // In this case we don't define what the state looks like, so anything goes, the router won't complain either
+    // In this case we don't define what the data looks like, so anything goes, the router won't complain either
     const uncheckedRoute = (ctx: Context) => {
-        console.log(ctx.state.doesntexist)
+        console.log(ctx.data.doesntexist)
     };
     router.get("/ex2", uncheckedRoute);
   });